feat(app): refactor modal with has
